// src/app/features/analytics/models/financial-summary.ts
export interface FinancialSummary {
  monthlyIncome: number;
  previousMonthlyIncome: number;
  monthlyExpense: number;
  previousMonthlyExpense: number;
  totalBalance: number;
  previousTotalBalance: number;
  incomeTransactionCount: number;
  expenseTransactionCount: number;
  incomeGroupsCount: number;
  expenseGroupsCount: number;
}